Group the EventListener(s) by domain per folder
