+++ src/Framework/EventListener/ClassResolver/ResolvedClassTryFormParentEvent.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gacela\Framework\EventListener\ClassResolver;
+
+use Gacela\Framework\EventListener\AbstractGacelaEvent;
+
+final class ResolvedClassTryFormParentEvent extends AbstractGacelaEvent
+{
+}
