--- src/Framework/EventListener/Event/ResolvedClassCachedEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gacela\Framework\EventListener\Event;
-
-final class ResolvedClassCachedEvent extends AbstractGacelaEvent
-{
-}
