--- src/Framework/EventListener/Event/ResolvedClassCreatedEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gacela\Framework\EventListener\Event;
-
-final class ResolvedClassCreatedEvent extends AbstractGacelaEvent
-{
-}
