--- src/Framework/EventListener/Event/ResolvedClassTryFormParentEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gacela\Framework\EventListener\Event;
-
-final class ResolvedClassTryFormParentEvent extends AbstractGacelaEvent
-{
-}
