--- src/Framework/EventListener/Event/ResolvedDefaultClassEvent.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Gacela\Framework\EventListener\Event;
-
-final class ResolvedDefaultClassEvent extends AbstractGacelaEvent
-{
-}
